Fixed bug with button
New instance of panelController was created in Presets.xib (instead of using existing one, which i was expecting)
To fix it i am setting file's owner of Presets.xib via initWithNibName: owner when create PresetsPanel object.
THere is no need for panelController class with such initialisation, so it was removed.